requirements.txt
torch
transformers
Pillow
tensorflow==2.15.0
numpy